Rounded days on service calculation

+++ backend/src/main/java/br/com/carreselling/application/service/model/VehicleSummary.java
@@ -16,5 +16,5 @@ public record VehicleSummary(UUID id,
                              BigDecimal servicesTotal,
                              BigDecimal totalCost,
                              String assignedPartnerName,
-                             int yardDays) {
+                             double yardDays) {
 }
